requirements.txt
torch==1.13.1
torchvision==0.14.1
torchaudio==0.13.1
transformers
numpy
matplotlib
pandas
tensorflow
keras
openpifpaf
python-dotenv
selenium
opencv-python
ultralytics
git+https://github.com/developer0hye/onepose.git